Move queue factory e2e tests to adapter suites
